\name{cdc__ref__data}
\alias{cdc__ref__data}
\docType{data}
\title{CDC Growth Chart Reference Data (2000)}
\description{
LMS parameters (L, M, S) for age- and sex-standardized growth metrics.
}
\usage{
data(cdc__ref__data)
}
\format{
A data.table with columns for sex, age, and LMS values.
}
\source{
\url{https://www.cdc.gov/growthcharts/percentile_data_files.htm}
}
\keyword{datasets}