\name{NHanes}
\alias{NHanes}
\docType{data}
\title{Sample Anthropometric Data from NHANES}
\description{
A sample dataset containing height, weight, sex, and age for testing the cdcanthro function.
}
\usage{
data(NHanes)
}
\format{
A data frame or data table with observations on multiple anthropometric variables.
}
\keyword{datasets}